updating model,  fixing aplication  and initializing frontend
