+++ backend/demo/src/main/java/com/VeterinarioBackendApplication.java
@@ -1,0 +1,13 @@
+package com;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class VeterinarioBackendApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(VeterinarioBackendApplication.class, args);
+	}
+
+}
