+++ backend/demo/src/main/java/com/dto/animal/AnimalSimpleDto.java
@@ -1,0 +1,14 @@
+package com.dto.animal;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class AnimalSimpleDto {
+    private int id;
+    private String nome;
+    private String especie;
+}
