+++ backend/demo/src/main/java/com/dto/cliente/ClienteSimpleDto.java
@@ -1,0 +1,14 @@
+package com.dto.cliente;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class ClienteSimpleDto {
+        private int cpf;
+        private String nome;
+
+    }
