+++ backend/demo/src/main/java/com/dto/medicamento/Medicamentodto.java
@@ -1,0 +1,21 @@
+package com.dto.medicamento;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import jakarta.validation.constraints.NotBlank;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class Medicamentodto {
+    @JsonProperty(access = JsonProperty.Access.READ_ONLY)
+    private long id;
+
+    @NotBlank(message = "Informe o nome do Medicamento")
+    private String nome;
+
+    private String descricao;
+
+}
