+++ backend/demo/src/main/java/com/dto/veterinario/VeterinarioSimpleDto.java
@@ -1,0 +1,13 @@
+package com.dto.veterinario;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class VeterinarioSimpleDto {
+    private int cpf;
+    private String nome;
+}
