+++ backend/demo/src/main/java/com/model/Cliente.java
@@ -1,0 +1,34 @@
+package com.model;
+
+import jakarta.persistence.*;
+import lombok.Data;
+
+import java.util.List;
+
+@Data
+@Entity
+@Table(name = "tb_cliente")
+public class Cliente {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private long id;
+
+    private int cpf;
+
+    private String nome;
+
+    private String email;
+
+    private String endereco;
+
+    private String telefone;
+
+    @ManyToMany (mappedBy = "cliente")
+    private List<Animal> animal;
+
+    @ManyToMany (mappedBy = "cliente")
+    private List<Veterinario> veterinario;
+
+
+}
