+++ backend/demo/src/main/java/com/model/Consulta.java
@@ -1,0 +1,32 @@
+package com.model;
+
+import jakarta.persistence.*;
+import lombok.Data;
+
+import java.time.LocalDateTime;
+import java.util.List;
+
+@Data
+@Entity
+@Table(name = "tb_consulta")
+public class Consulta {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private long id;
+
+    private String texto;
+
+    private LocalDateTime dataCriacao;
+
+    @OneToMany(mappedBy = "consulta", cascade = CascadeType.ALL)
+    private List<MedicamentoItem> medicamentoItem;
+
+    @ManyToOne
+    @JoinColumn(name = "animal_id")
+    private Animal animal;
+
+    @ManyToOne
+    @JoinColumn(name = "veterinario_id")
+    private Veterinario veterinario;
+
+}
