+++ backend/demo/src/main/java/com/model/Vacina.java
@@ -1,0 +1,26 @@
+package com.model;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+
+import java.time.LocalDateTime;
+
+
+@Entity
+@Table(name = "tb_vacina")
+public class Vacina {
+    @Id
+    @GeneratedValue(strategy = jakarta.persistence.GenerationType.IDENTITY)
+    private int id;
+
+    private String nome;
+
+    private String descricao;
+
+    private LocalDateTime dataAplicacao;
+
+    private LocalDateTime dataValidade;
+
+}
