--- backend/demo/src/main/java/com/veterinario/VeterinarioBackendApplication.java
@@ -1,13 +1,0 @@
-package com.veterinario;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class VeterinarioBackendApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(VeterinarioBackendApplication.class, args);
-	}
-
-}
