--- backend/demo/src/main/java/com/veterinario/dto/animal/AnimalSimpleDto.java
@@ -1,14 +1,0 @@
-package com.veterinario.dto.animal;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class AnimalSimpleDto {
-    private int id;
-    private String nome;
-    private String especie;
-}
