--- backend/demo/src/main/java/com/veterinario/dto/cliente/ClienteSimpleDto.java
@@ -1,14 +1,0 @@
-package com.veterinario.dto.cliente;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class ClienteSimpleDto {
-        private int cpf;
-        private String nome;
-
-    }
