--- backend/demo/src/main/java/com/veterinario/dto/medicamentos/Medicamentosdto.java
@@ -1,23 +1,0 @@
-package com.veterinario.dto.medicamentos;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import jakarta.validation.constraints.NotBlank;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class Medicamentosdto {
-    @JsonProperty(access = JsonProperty.Access.READ_ONLY)
-    private long id;
-
-    @NotBlank(message = "Informe o nome do Medicamento")
-    private String nome;
-
-    private int quantidade;
-
-    private String descricao;
-
-}
