--- backend/demo/src/main/java/com/veterinario/dto/veterinario/VeterinarioSimpleDto.java
@@ -1,13 +1,0 @@
-package com.veterinario.dto.veterinario;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class VeterinarioSimpleDto {
-    private int cpf;
-    private String nome;
-}
