--- backend/demo/src/main/java/com/veterinario/enums/Genero.java
@@ -1,16 +1,0 @@
-package com.veterinario.enums;
-
-import com.fasterxml.jackson.annotation.JsonValue;
-
-public enum Genero {
-    F("Feminino"),
-    M("Masculino");
-
-    final private String genero;
-
-    Genero(String genero) {this.genero = genero;}
-
-    @JsonValue
-    public String getGenero() {return genero;}
-
-}
