--- backend/demo/src/main/java/com/veterinario/model/Cliente.java
@@ -1,34 +1,0 @@
-package com.veterinario.model;
-
-import jakarta.persistence.*;
-import lombok.Data;
-
-import java.util.List;
-
-@Data
-@Entity
-@Table(name = "cliente")
-public class Cliente {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
-
-    private int cpf;
-
-    private String nome;
-
-    private String email;
-
-    private String endereco;
-
-    private String telefone;
-
-    @ManyToMany
-    private List<Animal> animais;
-
-    @ManyToMany
-    private List<Veterinario> veterinarios;
-
-
-}
