--- backend/demo/src/main/java/com/veterinario/model/Medicamento.java
@@ -1,19 +1,0 @@
-package com.veterinario.model;
-
-import jakarta.persistence.*;
-import lombok.Data;
-
-import java.util.List;
-@Data
-@Entity
-@Table(name = "medicamento")
-public class Medicamento {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
-
-    private String nome;
-
-    private String descricao;
-
-}
