--- backend/demo/src/main/java/com/veterinario/model/MedicamentoItem.java
@@ -1,22 +1,0 @@
-package com.veterinario.model;
-
-import jakarta.persistence.*;
-import lombok.Data;
-
-@Data
-@Entity
-@Table(name = "medicamento_item")
-public class MedicamentoItem {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
-
-    private String nome;
-
-    private String descricao;
-
-    private int quantidade;
-
-    @ManyToOne
-    private Consulta consulta;
-}
