+++ backend/demo/src/main/resources/application.properties
@@ -1,11 +1,10 @@
 spring.application.name=backend
 server.port=8080
 
-spring.datasource.url= jdbc:mysql//localhost33306/appveterinario
+spring.datasource.url= jdbc:mysql://localhost:3306/appveterinario
 spring.datasource.username= root
 spring.datasource.password= password
 
 spring.jpa.show-sql=true
 spring.jpa.generate-ddl=false
 spring.jpa.hibernate.ddl-auto= update
-spring.jpa.properties.hibernate.dialect= org.hibernate.dialect.MyDQL8Dialect
