+++ backend/observacoes.txt
@@ -1,0 +1,4 @@
+Olhar:
+Atualizar sql
+
+Tipo de loading no Model (LAZY, EAGER)
